Correction des 115 simulations 300 EUR: 164 montants, tous a la hausse
Ecart moyen +7.00 EUR (+2.03%), median +2.73, max +53.11, 164/164 en hausse.
Buy&Hold NDX 349.93 -> 352.39 EUR, exactement la valeur annoncee au #375.
Ampleur modeste car la fenetre est courte (63 seances): le terme d'erreur
croit avec l'horizon.

2 sims exclues et signalees: pead_sim_300e (rendements deja simples, cumprod
correct) et dollar_neutral_composite_vol_targeted (npz amont non verifie).

Campagne de correction terminee (#375->#381): 20 PASS tombes, 0 gagne.

+++ finance/trading/scripts/nonml_multimarket_breadth_vol_targeting_overlay_sim_300e.py
@@ -34,7 +34,9 @@ def main():
     bh_full_all = np.log(close_full[1:] / close_full[:-1])
 
     breadth = compute_multimarket_breadth_series(dates_idx)
-    breadth_raw = breadth.values
+    # astype(float) : sous pandas >= 3 la serie peut ressortir en dtype object,
+    # que np.isnan refuse. Conversion explicite, aucun changement de valeur.
+    breadth_raw = breadth.values.astype(float)
     gate_aligned = np.where(np.isnan(breadth_raw), False, breadth_raw >= BREADTH_THRESHOLD)
 
     pos_full_all = combined_position(bh_full_all, gate_aligned)
